moved pages, fixed npm errors
