+++ src/angular-app/languageforge/review-suggest/src/client/app/register/register.component.ts
